plot the traj. in Representation Space; tset_env.py
